Refactor so that EditItemForm displays correctly when edit button is clicked

+++ client/src/components/items/Items.js
@@ -1,5 +1,6 @@
 import React, { Component } from 'react'
 import Item from './Item'
+import EditItemForm from './EditItemForm'
 
 class Items extends Component {
   render() {
@@ -7,17 +8,30 @@ class Items extends Component {
       <div className="items">
         {this.props.items.map(item => {
           if (item.collection && this.props.collection && item.collection.id === this.props.collection.id) {
-            return (
-              <Item
-                key={item.id}
-                item={item}
-                deleteItem={this.props.deleteItem}
-                updateItem={this.props.updateItem}
-                editCollectionId={this.props.editCollectionId}
-                />
-            );
+            if (this.props.editItemId === item.id) {
+              return (
+                <EditItemForm
+                  item={item}
+                  collection={this.props.collection}
+                  key={item.id}
+                  updateItem={this.props.updateItem}
+                  toggleEditOff={this.props.toggleEditOff}
+                  />
+              );
+            } else {
+              return (
+                <Item
+                  key={item.id}
+                  item={item}
+                  deleteItem={this.props.deleteItem}
+                  toggleEditOn={this.props.toggleEditOn}
+                  editCollectionId={this.props.editCollectionId}
+                  />
+              );
+            }
           }
-        })}
+        return null;})
+        }
       </div>
     )
   }
